user restrictions

+++ src/components/PlantDelete/PlantDelete.js
@@ -2,7 +2,7 @@ import './PlantDelete.css';
 import React, {useContext, useEffect, useState} from 'react';
 import { useForm } from 'react-hook-form';
 import axios from "axios";
-import {BrowserRouter} from "react-router-dom";
+import {BrowserRouter, Link} from "react-router-dom";
 
 
 import {GrTrash} from "react-icons/gr";
@@ -12,21 +12,25 @@ function PlantDelete (props) {
     // nog doen: een waarschuwing!
     const [error, setError] = useState('');
     const [Success, toggleSuccess] = useState(false);
+    const [loading, toggleLoading] = useState(false);
     //const[contents] = useContext(ItemsContext);
    // console.log("props in itemDelete", props.id)
-
-
     // const { handleSubmit, formState: { errors }, register } = useForm();
     const itemId = props.id;
-    //console.log('itemId',itemId)
+
 
     async function deletePlantHandler () {
-        try {
-            await axios.delete(`http://localhost:8080/api/v1/plants/files/${itemId}`)
-            console.log('De plant is succesvol verwijderd')
+        // console.warn('Weet je het zeker?')
+        if(window.confirm("weet je het zeker?")){
+            toggleLoading(true);
+            try {
+                await axios.delete(`http://localhost:8080/api/v1/plants/files/${itemId}`)
+                // console.log('De plant is succesvol verwijderd')
+                toggleSuccess(true);
 
-        } catch (e) {
-            console.log('"het is niet gelukt, error: " '+ e)
+            } catch (e) {
+                console.log('het verwijderen is niet gelukt, error: '+ e)
+            }
         }
     }
 
@@ -34,18 +38,20 @@ function PlantDelete (props) {
 
     return (
         // <BrowserRouter forceRefresh>
-            <div >
+            <div >{!Success ?
+            <button
+                disabled={loading}
+                className='delete-plnt'
+                type='submit'
+                onClick={deletePlantHandler}
+                // onReset={reset}
+            >
+                verwijder
+                <GrTrash/>
+            </button>:<Link to={"/"}>Naar Overzicht</Link>
+            }
 
-                <button
-                    className='delete-plnt'
-                    type='submit'
-                    onClick={deletePlantHandler}
-                    // onReset={reset}
-               >
-                    verwijder
-                    <GrTrash/>
-
-                </button>
+                {Success &&<p>De plant is succesvol verwijderd</p>}
             </div>
         // </BrowserRouter>
     )
